Create pullcount.sh
Used to pull in the results from nightly cron outputs on all the connectors and come up with a total device count across the organization. Can also be run in a cron job or manually. Easy to run from ESM server if it has access to all the connectors via https.